requirements.txt
fastapi
uvicorn
psycopg2-binary
warcio
beautifulsoup4
lxml
pandas
rapidfuzz
requests
langchain-google-genai
langchain